+++ previous-versions/common/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.34.0] previous-versions/common
 ### Releases
 - [PyPI v0.34.0] https://pypi.org/project/codat-common/0.34.0 - previous-versions/common
+
+## 2023-10-16 01:25:31
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.99.2 (2.155.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.34.1] previous-versions/common
+### Releases
+- [PyPI v0.34.1] https://pypi.org/project/codat-common/0.34.1 - previous-versions/common